requirements.txt
blinker==1.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.3.2
fonttools==4.39.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lexicalrichness==0.5.0
MarkupSafe==2.1.2
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pdftotext==2.2.2
Pillow==9.5.0
pyparsing==3.0.9
pyphen==0.14.0
pyspellchecker==0.7.2
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
textblob==0.17.1
textstat==0.7.3
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
Werkzeug==2.3.4
